requirements.txt
# AeroStream Requirements
# =======================
# Wind tunnel data processing platform

# Database
pyodbc>=5.0.1

# Streaming
kafka-python>=2.0.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
pyarrow>=14.0.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Frontend
# NOTE: The current UI is a static HTML/D3 dashboard served from `frontend/`.
# Streamlit/Plotly were removed because no code in this repo imports them.

# Configuration
python-dotenv>=1.0.0

# Secret Management (OpenBao/Vault)
hvac>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
ruff>=0.1.0
